feat(#12)중복 코드 제거

+++ src/main/java/org/teamsai/saibackend/domain/account/dto/LinkableAccountResponse.java
@@ -2,12 +2,14 @@ package org.teamsai.saibackend.domain.account.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public record LinkableAccountResponse(
         Long accountId,
         String accountNumber,
         String accountName,
         String bankCode,
-        Long balance,
+        BigDecimal balance,
         String accountHolderName
 ) {
 }
